EVG-6496 Add autoupdate function to commands before property
