+++ cmd/evergreen/evergreen.go
@@ -1,8 +1,6 @@
 package main
 
 import (
-	"context"
-	"fmt"
 	"os"
 	"path/filepath"
 	"runtime"
@@ -17,7 +15,6 @@ import (
 	"github.com/mongodb/grip"
 	"github.com/mongodb/grip/level"
 	"github.com/mongodb/grip/send"
-	"github.com/pkg/errors"
 	"github.com/urfave/cli"
 )
 
@@ -103,18 +100,6 @@ func buildApp() *cli.App {
 	}
 
 	app.Before = func(c *cli.Context) error {
-		confPath := c.String("conf")
-		conf, err := operations.NewClientSettings(confPath)
-		if err != nil {
-			return errors.Wrap(err, "problem loading configuration")
-		}
-		if conf.AutoUpgradeCLI {
-			ctx, cancel := context.WithCancel(context.Background())
-			defer cancel()
-			if err = operations.CheckAndUpdateVersion(conf, ctx, true, false, true); err != nil {
-				fmt.Println("Automatic CLI update failed, continuing command execution")
-			}
-		}
 		return loggingSetup(app.Name, c.String("level"))
 	}
