+++ operations/export.go
@@ -61,6 +61,7 @@ func Export() cli.Command {
 				Usage: "write output in json, otherwise uses CSV",
 			}),
 		Before: mergeBeforeFuncs(
+			autoUpdateCLI,
 			requireStringFlag(statFlagName),
 			requireStringValueChoices(granularityFlagName,
 				[]string{granularityDays, granularityHours, granularityMinutes, granularitySeconds}),
