+++ operations/patch_finalize.go
@@ -13,7 +13,7 @@ func PatchFinalize() cli.Command {
 		Name:   "finalize-patch",
 		Usage:  "finalize an existing patch",
 		Flags:  addPatchIDFlag(),
-		Before: requirePatchIDFlag,
+		Before: mergeBeforeFuncs(autoUpdateCLI, requirePatchIDFlag),
 		Action: func(c *cli.Context) error {
 			confPath := c.Parent().String(confFlagName)
 			patchID := c.String(patchIDFlagName)
