+++ operations/patch_list.go
@@ -18,8 +18,9 @@ func PatchList() cli.Command {
 	)
 
 	return cli.Command{
-		Name:  "list-patches",
-		Usage: "show existing patches",
+		Name:   "list-patches",
+		Before: autoUpdateCLI,
+		Usage:  "show existing patches",
 		Flags: mergeFlagSlices(addPatchIDFlag(
 			cli.IntFlag{
 				Name:  joinFlagNames(numberFlagName, "n"),
