+++ operations/version.go
@@ -12,6 +12,7 @@ func Version() cli.Command {
 		Name:    "version",
 		Aliases: []string{"v"},
 		Usage:   "prints the revision of the current binary",
+		Before:  autoUpdateCLI,
 		Action: func(c *cli.Context) error {
 			fmt.Println(evergreen.ClientVersion)
 			return nil
